Create Access Control

+++ app/Http/Controllers/PostController.php
@@ -90,6 +90,9 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if (\auth()->user()->id !== $post->user_id) {
+            return \redirect('/posts')->with('error', 'Unauthorized Page');
+        }
         return \view('posts.edit', ['post' => $post]);
     }
 
@@ -125,6 +128,9 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if (\auth()->user()->id !== $post->user_id) {
+            return \redirect('/posts')->with('error', 'Unauthorized Page');
+        }
         $post->delete();
         return \redirect('/posts')->with('success' , 'Post Deleted');
     }
